improve performance of pustnuc and nucmer
Previously each `Node` in the alignment graph was wasting huge
amounts of memory. Each `Score` structure had seven bytes of padding
at the end. Thus, roughly a third of the memory occupied by a `Node` is
unused. With the new format, values are tightly packed making a `Node`
smaller. For an unrepresentative sample of two E. coli genomes, the new
postnuc is 10% faster and uses 50% less memory. The former might be due
to better caching.